Added support for multiple biomes

+++ Mesh Generation/Assets/Scripts/Spawnable.cs
@@ -15,7 +15,7 @@ public class Spawnable : ScriptableObject
     [Range(0, 100)]
     public int regionDensity;
     [Header("Structure Mask Settings")]
-    public float structureMaskOctaves = 3;
+    public int structureMaskOctaves = 3;
     public float structureMaskLacunarity = 2;
     public float structureMaskFrequency = 0.005f;
     public float structureMaskWeightedStrength = 1;
